pyproject.toml
[project]
name = "spamanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.121.3",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
]
